% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_materials_itewr.R
\name{get_materials_itewr}
\alias{get_materials_itewr}
\title{Download Book Materials}
\usage{
get_materials_itewr(dir = ".")
}
\arguments{
\item{dir}{A \code{character} specifying the path to the directory where the files will be saved. Defaults to the current working directory.}
}
\description{
This function allows you to download supplements for \emph{Introduction to Econometrics with R}, 
an interactive online-companion to the textbook \emph{Introduction to Econometrics} (Stock & Watson 2015). The book can be found \href{https://www.econometrics-with-r.org}{here}.
}
\details{
The function is a wrapper for \code{get_repo_files()} and \code{squeeze_rmds()}. 
It downloads materials such as datasets and R codes from the GitHub repository of the book (\emph{mca91/EconometricsWithR}) and saves them to the current working directory (or a location of choice provided to the argument \code{dir}) according to the following structure:
\itemize{
  \item \code{/ITEWR/Rmds} (.Rmd files)
  \item \code{/ITEWR/Data} (Datasets)
  \item \code{/ITEWR/Rcodes} (R Scripts for each chapter)
}
}
\section{References}{

Stock, J. H. and Watson, M. W. (2015). \emph{Introduction to Econometrics}, Third Update, Global Edition. Pearson Education Limited.
}

\examples{
# download materials
get_materials()
}
\keyword{download}
\keyword{supplements}
