% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_repo.R
\name{get_repo}
\alias{get_repo}
\title{Download a GitHub Repository}
\usage{
get_repo(repo = NULL, out = ".")
}
\arguments{
\item{repo}{\code{character} string. Name of the GitHub repository to be
downloaded. Must be specified in the form \code{"username/repository"}.}

\item{out}{A \code{character} string specifying the path to the output
directory where the repository files will be saved. Defaults to the current
working directory.}
}
\description{
This function allows to download a whole repository at GitHub.com. It is
intended as a convenience function for students working with
\emph{Introduction to Econometrics with R}, an interactive online-companion
to the textbook \emph{Introduction to Econometrics} (Stock & Watson 2015).
The companion can be found \href{https://www.econometrics-with-r.org}{here}.
}
\details{
The function downloads the zipped contents (\code{master.zip}) of
  the master branch of the GitHub repository supplied to the argument
  \code{repo} to the directory specified in \code{out} and unzips the archive
  afterwards in \code{<out>/Repository/}.
}
\section{References}{
 Stock, J. H. and Watson, M. W. (2015).
  \emph{Introduction to Econometrics}, Third Update, Global Edition. Pearson
  Education Limited.
}

\examples{
\dontrun{
# download the bookdown projekt repository
# "mca91/EconometricsWithR"
get_repo(repo = "mca91/EconometricsWithR")
}
}
\keyword{download}
\keyword{github}
\keyword{repository}
