% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_repo_files.R
\name{get_repo_files}
\alias{get_repo_files}
\title{Download Files From GitHub Repository}
\usage{
get_repo_files(repo = "mca91/EconometricsWithR", folder = "master",
  type = "any", out = ".")
}
\arguments{
\item{repo}{\code{character} A GitHub repository. Must be specified in the form \code{"username/repository"}.
Defaults to \code{"mca91/EconometricsWithR"}.}

\item{folder}{A \code{character} specifying the folder containing the files to be downloaded. Defaults to the \code{master} folder.}

\item{type}{A vector of type \code{character} specifying the file type(s) to be downloaded. All file types will be downloaded by default.}

\item{out}{A \code{character} specifying the path to the output directory where the files will be saved. Defaults to the current working directory.}
}
\description{
This function allows to download files from a specific folder located in a repository at GitHub.com. It is intended as a convenience function for students working with \emph{Introduction to Econometrics with R},
an interactive online-companion to the textbook \emph{Introduction to Econometrics} (Stock & Watson 2015). The companion can be found \href{https://www.econometrics-with-r.org}{here}. The function works with any GitHub repository.
}
\section{References}{

Stock, J. H. and Watson, M. W. (2015). \emph{Introduction to Econometrics}, Third Update, Global Edition. Pearson Education Limited.
}

\examples{
# download datasets from "mca91/EconometricsWithR"
get_repo_files(folder = "master/data")
}
\keyword{download}
\keyword{github}
