% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squeeze_rmds.R
\name{squeeze_rmds}
\alias{squeeze_rmds}
\title{A Simple Wrapper for Tangling Code in Bookdown Projects}
\usage{
squeeze_rmds(dir = ".", numbered = T, out = "Rcodes", prefix = "",
  ...)
}
\arguments{
\item{dir}{Path of the directory containing the \code{.Rmd} files to be tangled. Default is the current working directory.}

\item{numbered}{\code{logical}. If \code{TRUE}, only numbers in the original file names are retained.}

\item{out}{Path to the output directory where the tangled codes will be saved. Points the \code{./Rcode} by default.}

\item{prefix}{\code{character}. A prefix to be added to the \code{.R} file names.}

\item{...}{Further arguments passed to \code{knitr::purl()}.}
}
\description{
This function extracts R code from \code{.Rmd} files belonging to bookdown projects.
By default, it scans the working directory for \emph{numbered} \code{.Rmd} files and passes them to \code{knitr::purl()} which tangles the code chunks whithin to \code{.R} files.
It is intended as a convenience function for students working with \emph{Introduction to Econometrics with R},
an interactive online-companion to the textbook \emph{Introduction to Econometrics} (Stock & Watson 2015). The companion can be found \href{https://www.econometrics-with-r.org}{here}. The function works with arbitrary \code{.Rmd} files and is useful for working with other bookdown projects, too.
}
\examples{
\dontrun{
# tangle numbered chapters of a bookdown project at working directory
squeeze_rmds(numbered = TRUE, prefix = "Chapter_")
}
}
\keyword{bookdown}
\keyword{purl,}
