% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_repo_filenames.R
\name{get_repo_filenames}
\alias{get_repo_filenames}
\title{Get Filenames from GitHub Repository}
\usage{
get_repo_filenames(repo)
}
\arguments{
\item{repo}{A \code{character}. Path to a GitHub repository.}
}
\description{
This function allows to scrape filenames from a GitHub repository.
}
\examples{
# get filnames from 'EconometricsWithR/master/data'
get_repo_filenames(repo = "https://github.com/mca91/EconometricsWithR/tree/master/data")
}
\keyword{GitHub}
\keyword{filenames}
